+++ plugin/src/main/kotlin/ru/cian/rustore/publish/PluginConfig.kt
@@ -14,6 +14,8 @@ internal data class PluginConfig(
     val releaseNotes: List<ReleaseNotesConfig>?,
     val applicationId: String,
     val seoTags: List<SeoTag>,
+    val minAndroidVersion: String,
+    val developerContacts: DeveloperContactsConfig,
 )
 
 internal data class ReleasePhaseConfig(
@@ -28,4 +30,10 @@ internal data class Credentials(
 internal data class ReleaseNotesConfig(
     val lang: String,
     val newFeatures: String,
+)
+
+internal data class DeveloperContactsConfig(
+    val email: String,
+    val website: String? = null,
+    val vkCommunity: String? = null,
 )